Use custom operators to improve type check performance
By reusing standard Swift operators, we're exponentially increasing type
checker complexity. This can easily result in times of multiple seconds
to type check a file using SwiftAutoLayout, or expressions that are too
complex for the compiler to complete in time.

This change effectively introduces a "namespace" for SwiftAutoLayout
operators, by prefixing each operator with a "•".